refactor: Reorganize routes, more tests
